feat: rough ledstrip and serial defines

+++ lib/Networking_by_B5/Networking_by_B5.h
@@ -50,7 +50,7 @@ typedef struct // Maak een variabele InputData
 
     uint16_t NTC_RAW_DATA;
     uint16_t PRESSURE_RAW_DATA;
-    uint8_t HEARTBEAT_RAW_DATA;
+    uint16_t HEARTBEAT_RAW_DATA;
     uint16_t SATURATION_RAW_DATA;
 
     bool PriorityState;
